+++ libraries/Pliant/Languages/Pdl/PdlExpression.cs
@@ -28,7 +28,7 @@ namespace Pliant.Languages.Pdl
         int ComputeHashCode()
         {
             return HashCode.Compute(
-                NodeType.GetHashCode());
+                ((int)NodeType).GetHashCode());
         }
     }
 
@@ -59,7 +59,7 @@ namespace Pliant.Languages.Pdl
         int ComputeHashCode()
         {
             return HashCode.Compute(
-                NodeType.GetHashCode(),
+                ((int)NodeType).GetHashCode(),
                 Term.GetHashCode());
         }
 
@@ -99,7 +99,7 @@ namespace Pliant.Languages.Pdl
         int ComputeHashCode()
         {
             return HashCode.Compute(
-                NodeType.GetHashCode(),
+                ((int)NodeType).GetHashCode(),
                 Term.GetHashCode(),
                 Expression.GetHashCode());
         }
