+++ libraries/Pliant/Languages/Pdl/PdlLexerRuleExpression.cs
@@ -34,7 +34,7 @@ namespace Pliant.Languages.Pdl
         int ComputeHashCode()
         {
             return HashCode.Compute(
-                NodeType.GetHashCode(),
+                ((int)NodeType).GetHashCode(),
                 Term.GetHashCode());
         }
 
@@ -73,7 +73,7 @@ namespace Pliant.Languages.Pdl
         int ComputeHashCode()
         {
             return HashCode.Compute(
-                NodeType.GetHashCode(),
+                ((int)NodeType).GetHashCode(),
                 Term.GetHashCode(),
                 Expression.GetHashCode());
         }
